feat: support decimal sparepart qty (0.5 liter oil)
Sparepart stok sudah Numeric(15,2), tapi qty jual/beli masih INTEGER
sehingga pecahan tidak bisa dicatat: QtyControl membuang titik
("0.5" -> 5) dan formatNumber(0.5) memotong desimal jadi "0".

- migrasi f3a4b5c6d7e8: qty 2 tabel detail -> DECIMAL(15,2)
- schema/model qty part -> Decimal, jasa/servis tetap int
- QtyControl: satu titik desimal, decimal-pad, step 0.5 utk part
- formatQty baru (formatNumber utk uang tidak diubah)
- blok "Terpilih" dipindah ke atas katalog agar item terceklis
  selalu terlihat duluan saat edit dari queue

+++ backend/app/models/bengkel.py
@@ -138,7 +138,7 @@ class DetailPembelianSparePart(Base):
         ForeignKey("pembelian_spare_parts.id", ondelete="CASCADE")
     )
     spare_part_id: Mapped[int] = mapped_column(ForeignKey("spare_parts.id"))
-    qty: Mapped[int] = mapped_column(Integer, default=1)
+    qty: Mapped[Decimal] = mapped_column(Numeric(15, 2), default=1)
     harga_satuan: Mapped[Decimal] = mapped_column(Numeric(15, 2))
     subtotal: Mapped[Decimal] = mapped_column(Numeric(15, 2))
 
@@ -274,7 +274,7 @@ class DetailTransaksiSpareParts(Base):
         ForeignKey("transaksi_penjualan_bengkel.id", ondelete="CASCADE")
     )
     spare_part_id: Mapped[int] = mapped_column(ForeignKey("spare_parts.id"))
-    qty: Mapped[int] = mapped_column(Integer, default=1)
+    qty: Mapped[Decimal] = mapped_column(Numeric(15, 2), default=1)
     harga_beli: Mapped[Decimal] = mapped_column(Numeric(15, 2))  # HPP
     harga_jual: Mapped[Decimal] = mapped_column(Numeric(15, 2))
     subtotal: Mapped[Decimal] = mapped_column(Numeric(15, 2))
